requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio_client]
python-socketio[asgi]
google-genai
pygame
opencv-python
numpy